refactor: images/components names put in order
