+++ src/HeaderImage.js
@@ -1,8 +1,7 @@
-import React from "react";
 import styled from "styled-components";
 
 export const HeaderImage = styled.div`
-  background-image: url(${process.env.PUBLIC_URL + "/img/CoverImage.png"});
+  background-image: url(${process.env.PUBLIC_URL + "/img/header.png"});
   background-size: cover;
   min-height: 300px;
 `;
